+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0-beta.2 - TBD
 
 - Bugfix: Infinite digest loop when combining `am-time-ago` and `amTimezone` ([#178](https://github.com/urish/angular-moment/issues/178))
+- Bugfix: Cannot use angular-moment under webpack ([#108](https://github.com/urish/angular-moment/issues/108))
 
 ## 1.0.0-beta.1 - 2015-09-14